update 1. 将 redis key 常量移至 redis_conn.go 2. 修改 cookie 存放方式 3. 修改 test 用例
add
1. 用户退出接口新增 删除 redis 数据

+++ api/passport.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"errors"
 	"github.com/gin-gonic/gin"
 	"go-foodie-shop/cache"
 	"go-foodie-shop/serializer"
@@ -52,11 +53,16 @@ func UserLogin(c *gin.Context) {
 
 // UserLogout 用户退出
 func UserLogout(c *gin.Context) {
-	deleteCookie(c)
-	c.JSON(200, serializer.Response{
-		Status: 200,
-		Msg:    "登出成功",
-	})
+
+	userId, exist := c.GetQuery("userId")
+	if !exist {
+		c.JSON(400, ErrorResponse(errors.New("userId 不能为空")))
+	} else {
+		cache.RedisClient.Del(ShopCart + ":" + userId)
+		deleteCookie(c)
+		c.JSON(200, SuccessResponse(nil))
+	}
+
 }
 
 func deleteCookie(c *gin.Context) {
